+++ internal/pkg/impl/host/app/event_manager/manager_test.go
@@ -1,4 +1,4 @@
-package events
+package event_manager
 
 import (
 	"testing"
