+++ internal/pkg/impl/host/app/project_tooling_factory.go
@@ -2,13 +2,13 @@ package app
 
 import (
 	"github.com/spaulg/solo/internal/pkg/impl/host/app/context"
-	"github.com/spaulg/solo/internal/pkg/impl/host/app/events"
+	"github.com/spaulg/solo/internal/pkg/impl/host/app/event_manager"
 	"github.com/spaulg/solo/internal/pkg/impl/host/infra/container"
 )
 
 func ProjectToolingFactory(soloCtx *context.CliContext) (*ProjectTooling, error) {
 	// Event manager
-	eventManager := events.GetEventManagerInstance()
+	eventManager := event_manager.GetEventManagerInstance()
 
 	// Container orchestrator factory
 	orchestratorFactory := container.NewOrchestratorFactory(soloCtx, eventManager)
