+++ internal/pkg/impl/host/infra/container/progress/stream.go
@@ -7,7 +7,7 @@ import (
 	"io"
 
 	"github.com/spaulg/solo/internal/pkg/impl/host/app/context"
-	events_types "github.com/spaulg/solo/internal/pkg/types/host/app/events"
+	events_types "github.com/spaulg/solo/internal/pkg/impl/host/app/event_manager/events"
 )
 
 type ProgressEventStreamer struct {
