+++ test/mocks/host/app/audit/auditor.go
@@ -3,7 +3,7 @@ package audit
 import (
 	"github.com/stretchr/testify/mock"
 
-	"github.com/spaulg/solo/internal/pkg/types/host/app/events"
+	"github.com/spaulg/solo/internal/pkg/impl/host/app/event_manager/events"
 )
 
 type MockAuditor struct {
