+++ test/mocks/host/app/events/subscriber.go
@@ -3,7 +3,7 @@ package events
 import (
 	"github.com/stretchr/testify/mock"
 
-	events_types "github.com/spaulg/solo/internal/pkg/types/host/app/events"
+	events_types "github.com/spaulg/solo/internal/pkg/impl/host/app/event_manager/events"
 )
 
 type MockSubscriber struct {
