+++ test/mocks/host/app/wms/workflow_guard.go
@@ -4,7 +4,7 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	workflowcommon "github.com/spaulg/solo/internal/pkg/impl/common/domain/wms"
-	events_types "github.com/spaulg/solo/internal/pkg/types/host/app/events"
+	events_types "github.com/spaulg/solo/internal/pkg/impl/host/app/event_manager/events"
 )
 
 type MockWorkflowGuard struct {
